requirements.txt
django
django-environ
graphene-django
djangorestframework
dj-rest-auth
django-allauth
psycopg2-binary
pytest
pytest-django
wheel
requests
PyJWT
cryptography
pytest 
pytest-django
djangorestframework-simplejwt